Read me
This is a Python script that creates a graphical user interface (GUI) for an application called "Massilia Finance". The GUI is built using the PySide2 library. The application allows a user to input data (e.g. an expense amount, category, and date) and save it to an Excel file named "data.xlsx". The user can also delete data and view a history of the last 5 entries. The GUI has elements such as a logo, labels, dropdown menus, input fields, and buttons.